TP4/HIT2Nuevo/servidor/requirements.txt
flask
pika
opencv-python
numpy
redis
requests
google-cloud-storage